requirements.txt
setuptools
fastapi
uvicorn[standard]
asyncpg
pydantic
pydantic-settings
python-dotenv
passlib
bcrypt==3.2.0
python-jose
redis
